Re implement marshaller interfaces
