+++ format_test.go
@@ -146,4 +146,3 @@ func didFormatterPanic(pFormat string) bool {
 		return
 	}()
 }
-
